feat: add AbTestsServerModule
This module provides server-compatible version of the `cookieHandler` and `crawlerDetector` services. It optionally depends on
`REQUEST` from `@ngx-utils/cookies` and
and `CookiesService` from `@ngx-utils/cookies`. If they are not provided,
app will not crash. `cookieHandler` will fallback to a noop implementation in this case, and `crawlerDetector` will always detect user agent as non-crawler.